Test login

See merge request 21_HI007/21_hi007!22

+++ app/src/main/java/org/techtown/smim/database/gexercise.java
@@ -12,5 +12,5 @@ public class gexercise {
     public String ge_start_time;
     public String ge_end_time;
     public String video_url;
-    public int group_num;
+    public Long group_num;
 }
